make Julia 0.5 compatible (#2)
* make Julia 0.5 compatible

* add Julia 0.4 to travis test matrix

* make `get_header` work with UInt too

* huffmandata could be const

+++ src/table.jl
@@ -101,7 +101,7 @@ STATIC_TABLE =
      (b"www-authenticate", b"")
      ]
 
-function get_header(table::DynamicTable, index::Int)
+function get_header(table::DynamicTable, index)
     # IETF's table indexing is 1-based.
     if index <= length(STATIC_TABLE)
         return STATIC_TABLE[index]
